+++ example/test/reset_data_test.dart
@@ -4,7 +4,7 @@ import 'dart:io';
 import 'dart:async';
 import 'package:example/view_models/gantt_view_model.dart';
 import 'package:legacy_gantt_chart/legacy_gantt_chart.dart';
-import 'package:example/sync/websocket_gantt_sync_client.dart';
+// import 'package:example/sync/websocket_gantt_sync_client.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
 import 'package:stream_channel/stream_channel.dart';